Create Injection Sequence from MultiPlate (Multiple Plates)

Create Injection Sequence from MultiPlate (Multiple Plates)

Usage

# S4 method for class 'MultiPlate'
build_injec_seq(
  plate,
  method,
  repeat_std = 1,
  repeat_qc = 1,
  repeat_analyte = 1,
  blank_after_top_conc = TRUE,
  blank_at_end = TRUE,
  system_suitability = 0,
  blank_every_n = NULL,
  inject_vol,
  descr = "",
  prefix = Sys.Date(),
  suffix = "1",
  tray = 1,
  explore_mode = FALSE,
  conc_df = NULL
)

Arguments

plate

MultiPlate object

method

choose method from database

repeat_std

number of re-injections for calibration standards. Default is 1.

repeat_qc

number of re-injections for QC wells. Default is 1

repeat_analyte

number of re-injections for unknown samples. Default is 1

blank_after_top_conc

If TRUE, adding blank after high concentrations of standards and QCS.

blank_at_end

If True, adding blank at the end of queue.

system_suitability

Number of re-injections for suitability vial.

blank_every_n

If no QCs, frequency of injecting blanks between analytes.

inject_vol

volume of injection in micro liters.

descr

Run description.

prefix

string at the beginning of the filename. Default is today's date.

suffix

string to be added to the end of the filename. Default is "1".

tray

Location in sample manager.

explore_mode

options either TRUE or FALSE. Default if FALSE.

conc_df

data.frame matching compound name to a scaling factor. Maximum 20 compounds allowed.

Value

InjecListObj object