Read Chromatogram Files

This function reads chromatogram files from a directory and returns a data frame with the chromatogram data.

Usage

read_chrom(dir, format = "waters_raw", method)

Arguments

dir

directory for chromatraograms

method

LC-MS/MS method ID saved available in the database.

Examples

path <- system.file("extdata", "waters_raw_ex", package="PKbioanalysis")
main <- read_chrom(path)  
#> Error in checkmate::assertNumber(method, lower = 1, null.ok = FALSE): argument "method" is missing, with no default